Attempt at fixing 'thin start &' issue.
It has been observed that 'thin start &' terminates the intermediate shell in a way such that the Ruby process with CommandRunnerNG never sees an EOF on the stdout pipe. This meant we kept selecting until we timed out  - even though the shell exits immediately.

It kinda makes sense if the stdout fd for the intermediate shell is passed directly to the thin process. Then the pipe is never closed because thin now "owns" it, but the parent shell is dead, with thin now disowned.